requirements.txt
streamlit==0.90.0
tensorflow==2.7.0
Pillow==8.4.0
